moved round_to_decimals to Numeric so it also works for Fixnums

+++ lib/feldtruby/float.rb
@@ -1,6 +1,6 @@
 require 'feldtruby'
 
-class Float
+class Numeric
 	def round_to_decimals(numDecimals = 2)
 		factor = 10**numDecimals
 		(self * factor).round / factor.to_f
